+++ src/main/java/in/koreatech/koin/admin/shop/service/AdminShopService.java
@@ -1,10 +1,7 @@
 package in.koreatech.koin.admin.shop.service;
 
-import static in.koreatech.koin.domain.shop.model.review.ReportStatus.DELETED;
-
 import java.time.Clock;
 import java.time.LocalDate;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -12,17 +9,15 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import in.koreatech.koin.admin.shop.dto.menu.AdminCreateMenuCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.menu.AdminCreateMenuRequest;
-import in.koreatech.koin.admin.shop.dto.menu.AdminMenuCategoriesResponse;
-import in.koreatech.koin.admin.shop.dto.menu.AdminMenuDetailResponse;
-import in.koreatech.koin.admin.shop.dto.menu.AdminModifyMenuCategoryRequest;
-import in.koreatech.koin.admin.shop.dto.menu.AdminModifyMenuRequest;
-import in.koreatech.koin.admin.shop.dto.menu.AdminShopMenuResponse;
-import in.koreatech.koin.admin.shop.dto.review.AdminModifyShopReviewReportStatusRequest;
-import in.koreatech.koin.admin.shop.dto.review.AdminShopsReviewsResponse;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopCategoryRequest;
 import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest;
+import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest.InnerShopOpen;
 import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoriesOrderRequest;
 import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopCategoryRequest;
 import in.koreatech.koin.admin.shop.dto.shop.AdminModifyShopRequest;
@@ -30,35 +25,21 @@ import in.koreatech.koin.admin.shop.dto.shop.AdminShopCategoryResponse;
 import in.koreatech.koin.admin.shop.dto.shop.AdminShopParentCategoryResponse;
 import in.koreatech.koin.admin.shop.dto.shop.AdminShopResponse;
 import in.koreatech.koin.admin.shop.dto.shop.AdminShopsResponse;
-import in.koreatech.koin.admin.shop.exception.ShopCategoryNotEmptyException;
-import in.koreatech.koin.admin.shop.exception.ShopCategoryIllegalArgumentException;
-import in.koreatech.koin.admin.shop.repository.*;
-import in.koreatech.koin.domain.shop.exception.ReviewNotFoundException;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import in.koreatech.koin.admin.shop.dto.shop.AdminCreateShopRequest.InnerShopOpen;
-import in.koreatech.koin.admin.shop.dto.menu.AdminModifyMenuRequest.InnerOptionPrice;
 import in.koreatech.koin.admin.shop.exception.ShopCategoryDuplicationException;
-import in.koreatech.koin.domain.shop.model.menu.Menu;
+import in.koreatech.koin.admin.shop.exception.ShopCategoryIllegalArgumentException;
+import in.koreatech.koin.admin.shop.exception.ShopCategoryNotEmptyException;
+import in.koreatech.koin.admin.shop.repository.AdminEventArticleRepository;
+import in.koreatech.koin.admin.shop.repository.AdminShopCategoryMapRepository;
+import in.koreatech.koin.admin.shop.repository.AdminShopCategoryRepository;
+import in.koreatech.koin.admin.shop.repository.AdminShopParentCategoryRepository;
+import in.koreatech.koin.admin.shop.repository.AdminShopRepository;
 import in.koreatech.koin.domain.shop.model.menu.MenuCategory;
-import in.koreatech.koin.domain.shop.model.menu.MenuCategoryMap;
-import in.koreatech.koin.domain.shop.model.menu.MenuImage;
-import in.koreatech.koin.domain.shop.model.menu.MenuOption;
-import in.koreatech.koin.domain.shop.model.review.ReportStatus;
-import in.koreatech.koin.domain.shop.model.review.ShopReview;
-import in.koreatech.koin.domain.shop.model.review.ShopReviewReport;
 import in.koreatech.koin.domain.shop.model.shop.Shop;
 import in.koreatech.koin.domain.shop.model.shop.ShopCategory;
 import in.koreatech.koin.domain.shop.model.shop.ShopCategoryMap;
 import in.koreatech.koin.domain.shop.model.shop.ShopImage;
 import in.koreatech.koin.domain.shop.model.shop.ShopOpen;
 import in.koreatech.koin.domain.shop.model.shop.ShopParentCategory;
-import in.koreatech.koin.global.exception.KoinIllegalArgumentException;
 import in.koreatech.koin.global.model.Criteria;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
@@ -71,20 +52,16 @@ public class AdminShopService {
     private final Clock clock;
     private final EntityManager entityManager;
     private final AdminEventArticleRepository adminEventArticleRepository;
-    private final AdminMenuCategoryRepository adminMenuCategoryRepository;
     private final AdminShopCategoryMapRepository adminShopCategoryMapRepository;
     private final AdminShopCategoryRepository adminShopCategoryRepository;
     private final AdminShopParentCategoryRepository adminShopParentCategoryRepository;
     private final AdminShopRepository adminShopRepository;
-    private final AdminMenuRepository adminMenuRepository;
-    private final AdminShopReviewRepository adminShopReviewRepository;
-    private final AdminShopReviewCustomRepository adminShopReviewCustomRepository;
 
     public AdminShopsResponse getShops(Integer page, Integer limit, Boolean isDeleted) {
         Integer total = adminShopRepository.countAllByIsDeleted(isDeleted);
         Criteria criteria = Criteria.of(page, limit, total);
-        PageRequest pageRequest = PageRequest.of(criteria.getPage(), criteria.getLimit(),
-            Sort.by(Sort.Direction.ASC, "id"));
+        Sort sort = Sort.by(Sort.Direction.ASC, "id");
+        PageRequest pageRequest = PageRequest.of(criteria.getPage(), criteria.getLimit(), sort);
         Page<Shop> result = adminShopRepository.findAllByIsDeleted(isDeleted, pageRequest);
         return AdminShopsResponse.of(result, criteria);
     }
@@ -112,30 +89,6 @@ public class AdminShopService {
         return shopParentCategories.stream()
             .map(AdminShopParentCategoryResponse::from)
             .toList();
-    }
-
-    public AdminShopMenuResponse getAllMenus(Integer shopId) {
-        Shop shop = adminShopRepository.getById(shopId);
-        List<MenuCategory> menuCategories = adminMenuCategoryRepository.findAllByShopId(shop.getId());
-        Collections.sort(menuCategories);
-        return AdminShopMenuResponse.from(menuCategories);
-    }
-
-    public AdminMenuCategoriesResponse getAllMenuCategories(Integer shopId) {
-        Shop shop = adminShopRepository.getById(shopId);
-        List<MenuCategory> menuCategories = adminMenuCategoryRepository.findAllByShopId(shop.getId());
-        Collections.sort(menuCategories);
-        return AdminMenuCategoriesResponse.from(menuCategories);
-    }
-
-    public AdminMenuDetailResponse getMenu(Integer shopId, Integer menuId) {
-        adminShopRepository.getById(shopId);
-        Menu menu = adminMenuRepository.getById(menuId);
-        List<MenuCategory> menuCategories = menu.getMenuCategoryMaps()
-            .stream()
-            .map(MenuCategoryMap::getMenuCategory)
-            .toList();
-        return AdminMenuDetailResponse.createMenuDetailResponse(menu, menuCategories);
     }
 
     @Transactional
@@ -191,55 +144,6 @@ public class AdminShopService {
     }
 
     @Transactional
-    public void createMenu(Integer shopId, AdminCreateMenuRequest adminCreateMenuRequest) {
-        Shop shop = adminShopRepository.getById(shopId);
-        Menu menu = adminCreateMenuRequest.toEntity(shop);
-        Menu savedMenu = adminMenuRepository.save(menu);
-        for (Integer categoryId : adminCreateMenuRequest.categoryIds()) {
-            MenuCategory menuCategory = adminMenuCategoryRepository.getById(categoryId);
-            MenuCategoryMap menuCategoryMap = MenuCategoryMap.builder()
-                .menuCategory(menuCategory)
-                .menu(savedMenu)
-                .build();
-            savedMenu.getMenuCategoryMaps().add(menuCategoryMap);
-        }
-        for (String imageUrl : adminCreateMenuRequest.imageUrls()) {
-            MenuImage menuImage = MenuImage.builder()
-                .imageUrl(imageUrl)
-                .menu(savedMenu)
-                .build();
-            savedMenu.getMenuImages().add(menuImage);
-        }
-        if (adminCreateMenuRequest.optionPrices() == null) {
-            MenuOption menuOption = MenuOption.builder()
-                .option(savedMenu.getName())
-                .price(adminCreateMenuRequest.singlePrice())
-                .menu(menu)
-                .build();
-            savedMenu.getMenuOptions().add(menuOption);
-        } else {
-            for (var option : adminCreateMenuRequest.optionPrices()) {
-                MenuOption menuOption = MenuOption.builder()
-                    .option(option.option())
-                    .price(option.price())
-                    .menu(menu)
-                    .build();
-                savedMenu.getMenuOptions().add(menuOption);
-            }
-        }
-    }
-
-    @Transactional
-    public void createMenuCategory(Integer shopId, AdminCreateMenuCategoryRequest adminCreateMenuCategoryRequest) {
-        Shop shop = adminShopRepository.getById(shopId);
-        MenuCategory menuCategory = MenuCategory.builder()
-            .shop(shop)
-            .name(adminCreateMenuCategoryRequest.name())
-            .build();
-        adminMenuCategoryRepository.save(menuCategory);
-    }
-
-    @Transactional
     public void cancelShopDelete(Integer shopId) {
         Optional<Shop> shop = adminShopRepository.findDeletedShopById(shopId);
         if (shop.isPresent()) {
@@ -286,12 +190,6 @@ public class AdminShopService {
         );
     }
 
-    private void validateExistCategoryName(String name, Integer categoryId) {
-        if (adminShopCategoryRepository.existsByNameAndIdNot(name, categoryId)) {
-            throw ShopCategoryDuplicationException.withDetail("name: " + name);
-        }
-    }
-
     @Transactional
     public void modifyShopCategoriesOrder(AdminModifyShopCategoriesOrderRequest adminModifyShopCategoriesOrderRequest) {
         Map<Integer, ShopCategory> shopCategoryMap = adminShopCategoryRepository.findAll().stream()
@@ -309,32 +207,6 @@ public class AdminShopService {
     }
 
     @Transactional
-    public void modifyMenuCategory(Integer shopId, AdminModifyMenuCategoryRequest adminModifyMenuCategoryRequest) {
-        adminShopRepository.getById(shopId);
-        MenuCategory menuCategory = adminMenuCategoryRepository.getById(adminModifyMenuCategoryRequest.id());
-        menuCategory.modifyName(adminModifyMenuCategoryRequest.name());
-    }
-
-    @Transactional
-    public void modifyMenu(Integer shopId, Integer menuId, AdminModifyMenuRequest adminModifyMenuRequest) {
-        Menu menu = adminMenuRepository.getById(menuId);
-        adminShopRepository.getById(shopId);
-        menu.modifyMenu(
-            adminModifyMenuRequest.name(),
-            adminModifyMenuRequest.description()
-        );
-        menu.modifyMenuImages(adminModifyMenuRequest.imageUrls(), entityManager);
-        menu.modifyMenuCategories(adminMenuCategoryRepository.findAllByIdIn(adminModifyMenuRequest.categoryIds()),
-            entityManager);
-        if (adminModifyMenuRequest.isSingle()) {
-            menu.adminModifyMenuSingleOptions(adminModifyMenuRequest, entityManager);
-        } else {
-            List<InnerOptionPrice> optionPrices = adminModifyMenuRequest.optionPrices();
-            menu.adminModifyMenuMultipleOptions(optionPrices, entityManager);
-        }
-    }
-
-    @Transactional
     public void deleteShop(Integer shopId) {
         Shop shop = adminShopRepository.getById(shopId);
         shop.delete();
@@ -348,75 +220,13 @@ public class AdminShopService {
         adminShopCategoryRepository.deleteById(categoryId);
     }
 
+    private void validateExistCategoryName(String name, Integer categoryId) {
+        if (adminShopCategoryRepository.existsByNameAndIdNot(name, categoryId)) {
+            throw ShopCategoryDuplicationException.withDetail("중복되는 상점 카테고리명이 있습니다.: " + name);
+        }
+    }
+
     private boolean hasShops(Integer categoryId) {
         return adminShopCategoryMapRepository.existsByShopCategoryId(categoryId);
-    }
-
-    @Transactional
-    public void deleteMenuCategory(Integer shopId, Integer categoryId) {
-        MenuCategory menuCategory = adminMenuCategoryRepository.getById(categoryId);
-        if (!Objects.equals(menuCategory.getShop().getId(), shopId)) {
-            throw new KoinIllegalArgumentException("해당 상점의 카테고리가 아닙니다.");
-        }
-        adminMenuCategoryRepository.deleteById(categoryId);
-    }
-
-    @Transactional
-    public void deleteMenu(Integer shopId, Integer menuId) {
-        Menu menu = adminMenuRepository.getById(menuId);
-        if (!Objects.equals(menu.getShop().getId(), shopId)) {
-            throw new KoinIllegalArgumentException("해당 상점의 카테고리가 아닙니다.");
-        }
-        adminMenuRepository.deleteById(menuId);
-    }
-
-    @Transactional(readOnly = true)
-    public AdminShopsReviewsResponse getReviews(
-        Integer page,
-        Integer limit,
-        Boolean isReported,
-        Boolean hasUnhandledReport,
-        Integer shopId
-    ) {
-        Long reviewCount = adminShopReviewCustomRepository.countShopReview(shopId, isReported, hasUnhandledReport);
-        Criteria criteria = Criteria.of(page, limit, reviewCount.intValue());
-        PageRequest pageRequest = PageRequest.of(
-            criteria.getPage(),
-            criteria.getLimit(),
-            Sort.by(Sort.Direction.ASC, "id")
-        );
-
-        Page<ShopReview> reviews = adminShopReviewCustomRepository.findShopReview(
-            shopId,
-            isReported,
-            hasUnhandledReport,
-            pageRequest
-        );
-        return AdminShopsReviewsResponse.of(reviews, criteria);
-    }
-
-    @Transactional
-    public void modifyShopReviewReportStatus(
-        Integer reviewId,
-        AdminModifyShopReviewReportStatusRequest adminModifyShopReviewReportStatusRequest
-    ) {
-        ShopReview shopReview = adminShopReviewRepository.findById(reviewId)
-            .orElseThrow(() -> ReviewNotFoundException.withDetail("reviewId: " + reviewId));
-
-        List<ShopReviewReport> unhandledReports = shopReview.getReports().stream()
-            .filter(report -> report.getReportStatus() == ReportStatus.UNHANDLED)
-            .toList();
-
-        unhandledReports.forEach(
-            report -> report.modifyReportStatus(adminModifyShopReviewReportStatusRequest.reportStatus()));
-    }
-
-    @Transactional
-    public void deleteShopReview(Integer reviewId) {
-        ShopReview shopReview = adminShopReviewRepository.findById(reviewId)
-            .orElseThrow(() -> ReviewNotFoundException.withDetail("reviewId: " + reviewId));
-
-        shopReview.getReports().forEach(report -> report.modifyReportStatus(DELETED));
-        shopReview.deleteReview();
     }
 }
